+++ src/Components/AlterOperation.php
@@ -13,6 +13,8 @@ use PhpMyAdmin\SqlParser\Token;
 use PhpMyAdmin\SqlParser\TokensList;
 use function array_key_exists;
 use function in_array;
+use function is_numeric;
+use function is_string;
 
 /**
  * Parses an alter operation.
@@ -262,6 +264,12 @@ class AlterOperation extends Component
 
                 $state = 2;
             } elseif ($state === 2) {
+                $array_key = '';
+                if (is_string($token->value) || is_numeric($token->value)) {
+                    $array_key = $token->value;
+                } else {
+                    $array_key = $token->token;
+                }
                 if ($token->type === Token::TYPE_OPERATOR) {
                     if ($token->value === '(') {
                         ++$brackets;
@@ -281,9 +289,9 @@ class AlterOperation extends Component
                         );
                         break;
                     }
-                } elseif ((array_key_exists($token->value, self::$DB_OPTIONS)
-                    || array_key_exists($token->value, self::$TABLE_OPTIONS))
-                    && ! self::checkIfColumnDefinitionKeyword($token->value)
+                } elseif ((array_key_exists($array_key, self::$DB_OPTIONS)
+                    || array_key_exists($array_key, self::$TABLE_OPTIONS))
+                    && ! self::checkIfColumnDefinitionKeyword($array_key)
                 ) {
                     // This alter operation has finished, which means a comma
                     // was missing before start of new alter operation